add comment

+++ januscloud/proxy/dao/rd_room_dao.py
@@ -38,6 +38,10 @@ class RDRoomDao(object):
         room_key_list = []
         for room in room_list:
             room_key_list.append(self._key_room(room.room_id))
+
+        if len(room_key_list) == 0 
+            return None
+
         self._redis_client.delete(*room_key_list)
 
     def add(self, room):
